Create a separate iso handler from the image handler
The image handler will be the entry point for paths rooted at `/images/`
but we will need serve both isos and initrds from there so the existing
images handler will act as a router to those two sub-paths.
